+++ apps/example-expo/app/example-height-budget-scroll-view.tsx
@@ -1,6 +1,5 @@
 import { Fragment, useState } from 'react'
 import { Button, StyleSheet, Text, View } from 'react-native'
-import { useSharedValue } from 'react-native-reanimated'
 import {
   Backdrop,
   BottomSheet,
@@ -22,7 +21,6 @@ export default function ExampleHeightBudgetScrollView() {
   // When using dynamic sizing, it's important to set a max height
   // to prevent content taking up the entire screen
   const maxHeight = 600
-  const maxHeightShared = useSharedValue(maxHeight)
 
   // MARK: Renderers
 
@@ -57,7 +55,7 @@ export default function ExampleHeightBudgetScrollView() {
           <BottomSheetPresenter>
             <BottomSheetProvider>
               <BottomSheet styles={{ root: { maxHeight } }}>
-                <HeightBudgetProvider maxHeight={maxHeightShared}>
+                <HeightBudgetProvider maxHeight={maxHeight}>
                   <HeightClaim>
                     <BottomSheetHandle />
                   </HeightClaim>