+++ packages/react-native-the-sheet/src/core/bottom-sheet/bottom-sheet-flat-list.tsx
@@ -1,4 +1,7 @@
-import Animated, { type AnimatedRef } from 'react-native-reanimated'
+import Animated, {
+  useAnimatedStyle,
+  type AnimatedRef,
+} from 'react-native-reanimated'
 import { Gesture, GestureDetector } from 'react-native-gesture-handler'
 import type { BottomSheetFlatListProps } from './types'
 import { StyleSheet } from 'react-native'
@@ -6,9 +9,10 @@ import { useMemo } from 'react'
 import { useBottomSheetScrollViewUtils } from './hooks/use-bottom-sheet-scroll-view-utils'
 import { useBottomSheetPanGesture } from './hooks/use-bottom-sheet-pan-gesture'
 import { useBottomSheet } from './bottom-sheet-provider'
+import { useToSharedValue } from '../private/hooks/use-to-shared-value'
 
 export function BottomSheetFlatList<T>({
-  fill,
+  fill: propFill = false,
 
   onLayout: propOnLayout,
   onTouchStart: propOnTouchStart,
@@ -29,6 +33,8 @@ export function BottomSheetFlatList<T>({
 
   const getPanGesture = useBottomSheetPanGesture()
 
+  const fill = useToSharedValue(propFill)
+
   const panGesture = useMemo(() => {
     return getPanGesture()
   }, [getPanGesture])
@@ -45,6 +51,12 @@ export function BottomSheetFlatList<T>({
       onMomentumEnd: propOnMomentumEnd,
     })
 
+  const animatedStyle = useAnimatedStyle(() => {
+    return {
+      ...(fill.value ? styles.fill : undefined),
+    }
+  })
+
   return (
     <GestureDetector
       gesture={Gesture.Simultaneous(panGesture, Gesture.Native())}
@@ -52,7 +64,7 @@ export function BottomSheetFlatList<T>({
       <Animated.FlatList
         {...rest}
         ref={scrollViewRef as AnimatedRef<Animated.FlatList>}
-        style={[styles.root, fill && styles.fill, style]}
+        style={[styles.root, style, animatedStyle]}
         contentContainerStyle={contentContainerStyle}
         bounces={false} // iOS bounce ruins the scrollY <= 0 check
         onLayout={onLayout}
@@ -63,6 +75,8 @@ export function BottomSheetFlatList<T>({
     </GestureDetector>
   )
 }
+
+// MARK: Styles
 
 const styles = StyleSheet.create({
   fill: {