+++ packages/react-native-the-sheet/src/core/private/types.ts
@@ -1,0 +1,1 @@
+export type AnimatedProp<T> = T | { value: T }